src/main/java/ru/netology/taskmanager/Epic.java
package ru.netology.taskmanager;

public class Epic extends Task {

    public static String getSubtasks;
    protected String[] subtasks;

    public Epic(int id, String[] subtasks) {
        super(id); // вызов родительского конструктора
        this.subtasks = subtasks; // заполнение своих полей
    }

    public String[] getSubtasks() {
        return subtasks;
    }

    @Override
    public boolean matches(String query) {
        if (query == null || query.isEmpty() || subtasks == null) {
            return false;
        }
        for (String subtask : subtasks) {
            if (subtask.contains(query)) {
                return true;
            }
        }
        return false;
    }
}
